BF: Use session.auth as an attribute, do not try to call it as a method
In https://github.com/datalad/datalad-xnat/issues/8#issuecomment-929870134
I reported that an authentication attempt resulted in a TypeError.
I believe this is due to a change that happened in #33, which
implemented session.auth, but tried to call the attribute as a method.
This commit introduces the correct format, I believe.