ENH: Add a logging info on starting the authentication process
My rational to add this is UX-based: When running "datalad xnat-init"
it can take a while for the command to finish the query. Without any
change to the log level, there is nothing that the user is seeing -
the command just sits there and does not return until the query is
finished. Notifying the user about the process of authentication is
a quick feedback that there indeed is something happening.